Move version check files into utils
